Create get_student_list.py
Gets a list of students enrolled in a given course with a specific course id.  I just hardcoded the id to make life easier, but if you wanted to do something similar for all courses, you could just create a separate list with the appropriate scope in another file and then call that function and run the get-students function against the procured numbers list.

I'm certain that there is a better way to do all of this.  I am just struggling through as I go.