Send email from gmail account
This was a pain in the neck to figure out.  I had hoped that there would be a complete solution out on them interwebs, but there wasn't...at least, not one that I could understand.  So I pulled this together and it seems to work just fine.  Once I learn more about constructing mime objects I'll try to use this to automate sending out progress report emails.