Create create_message.py
Isolated function to create an html message to pass to the send_msg function from send_pr.  send_msg is invoked here since it was easier to just pull the data from individual records.  One day I'll go back to this project and have the data put into a dictionary or database to simplify the task of extracting the information needed to build the mime text.  This is definitely a hack job, but it works - so there's that.