Create create_pr.py
A complete reworking of the create_reports.py script.  Rather than dumping the data into a CSV for a mail merge, this function loads all of the desired data into a list so that it can be passed along to the mail functions.  This is u-g-l-y, but it works..so there's that.